+++ esm/index.js
@@ -1,8 +1,0 @@
-/* eslint-disable node/no-unpublished-import, node/no-unsupported-features/es-syntax, import/extensions */
-import UttoriWiki from '../src/wiki.js';
-
-export default UttoriWiki;
-
-export { default as config } from '../src/config.js';
-export { default as middleware } from '../src/middleware.js';
-export { default as UttoriWiki } from '../src/wiki.js';