+++ src/index.js
@@ -1,5 +1,5 @@
-module.exports = {
-  config: require('./config'),
-  wiki: require('./middleware'),
-  UttoriWiki: require('./wiki'),
-};
+export { default as config } from './config.js';
+export { default as wiki } from './middleware.js';
+export { default as UttoriWiki } from './wiki.js';
+export { default as wikiFlash } from './wiki-flash.js';
+export { default as AddQueryOutputToViewModel } from './plugins/add-query-output.js';