+++ src/middleware.js
@@ -1,17 +1,18 @@
-/** @type {Function} */
-let debug = () => {}; try { debug = require('debug')('Uttori.Wiki'); } catch {}
-const express = require('express');
+import express from 'express';
 
-const UttoriWiki = require('./wiki');
-const { middleware: flash } = require('./wiki-flash');
-const { UttoriWikiConfig } = require('./config');
+import UttoriWiki from './wiki.js';
+import { middleware as flash } from './wiki-flash.js';
+
+let debug = (..._) => {};
+/* c8 ignore next */
+try { const { default: d } = await import('debug'); debug = d('Uttori.middleware'); } catch {}
 
 /**
  *
- * @param {UttoriWikiConfig} config The UttoriWikiConfig to use.
+ * @param {import('./config.js').UttoriWikiConfig} config The UttoriWikiConfig to use.
  * @returns {express.Application} The Express server setup with UttoriWiki.
  */
-module.exports = function middleware(config) {
+function middleware(config) {
   const app = express();
 
   // Add flash messages if Express Session is avaliable.
@@ -22,7 +23,7 @@ module.exports = function middleware(config) {
     // console.log('config.middleware', config.middleware);
     for (const item of config.middleware) {
       const fn = item.shift();
-      if (fn && app[fn] && typeof app[fn] === 'function') {
+      if (fn && typeof fn === 'string' && app[fn] && typeof app[fn] === 'function') {
         debug(`app.${fn}(${JSON.stringify(item)})`);
         app[fn].call(app, ...item);
       }
@@ -31,7 +32,7 @@ module.exports = function middleware(config) {
 
   // https://expressjs.com/en/4x/api.html#express.static
   // https://webhint.io/docs/user-guide/hints/hint-http-cache/#examples-that-pass-the-hint
-  app.use(express.static(config.public_dir, {
+  app.use(express.static(config.publicPath, {
     etag: true, // Just being explicit about the default.
     lastModified: true, // Just being explicit about the default.
     immutable: true,
@@ -42,4 +43,6 @@ module.exports = function middleware(config) {
   const _wiki = new UttoriWiki(config, app);
 
   return app;
-};
+}
+
+export default middleware;