+++ test/create.test.js
@@ -1,10 +1,10 @@
-const test = require('ava');
-const sinon = require('sinon');
-const request = require('supertest');
+import test from 'ava';
 
-const { UttoriWiki } = require('../src');
+import sinon from 'sinon';
+import request from 'supertest';
+import { UttoriWiki } from '../src/index.js';
 
-const { config, serverSetup, seed } = require('./_helpers/server');
+import { config, serverSetup, seed } from './_helpers/server.js';
 
 test('create(request, response, _next): renders', async (t) => {
   t.plan(3);
@@ -16,7 +16,7 @@ test('create(request, response, _next): renders', async (t) => {
   t.is(response.status, 200);
   t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
   const title = response.text.match(/<title>(.*?)<\/title>/i);
-  t.is(title[1], 'New Document | Wiki');
+  t.is(title[1], 'New Document');
 });
 
 test('create(request, response, _next): can be replaced', async (t) => {
@@ -63,5 +63,5 @@ test('falls to 404 when miss matched key', async (t) => {
   t.is(express_response.status, 404);
   t.is(express_response.text.slice(0, 15), '<!DOCTYPE html>');
   const title = express_response.text.match(/<title>(.*?)<\/title>/i);
-  t.is(title[1], '404 Not Found | Wiki');
+  t.is(title[1], '404 Not Found');
 });