+++ test/historyDetail.test.js
@@ -1,11 +1,10 @@
-// @ts-nocheck
-const test = require('ava');
-const request = require('supertest');
-const sinon = require('sinon');
+import test from 'ava';
+import request from 'supertest';
+import sinon from 'sinon';
 
-const { UttoriWiki } = require('../src');
+import { UttoriWiki } from '../src/index.js';
 
-const { config, serverSetup, seed } = require('./_helpers/server');
+import { config, serverSetup, seed } from './_helpers/server.js';
 
 const response = { set: () => {}, redirect: () => {}, render: () => {} };
 
@@ -58,11 +57,11 @@ test('can be replaced', async (t) => {
   t.is(spy.called, true);
 });
 
-test('falls through to next when public_history is false', async (t) => {
+test('falls through to next when publicHistory is false', async (t) => {
   t.plan(1);
 
   const server = serverSetup();
-  const uttori = new UttoriWiki({ ...config, public_history: false }, server);
+  const uttori = new UttoriWiki({ ...config, publicHistory: false }, server);
   await seed(uttori);
   const [history] = await uttori.hooks.fetch('storage-get-history', 'demo-title', uttori);
   const express_response = await request(server).get(`/demo-title/history/${history[0]}`);