+++ test/middleware.test.js
@@ -1,26 +1,19 @@
-// @ts-nocheck
-const test = require('ava');
-const express = require('express');
-const defaultConfig = require('../src/config');
-const wiki = require('../src/middleware');
+import test from 'ava';
+
+import express from 'express';
+import defaultConfig from '../src/config.js';
+import wiki from '../src/middleware.js';
 
 test('wiki can run', (t) => {
   const config = {
     ...defaultConfig,
-    site_sections: [
-      {
-        title: 'Example',
-        description: 'Example description text.',
-        tag: 'example',
-      },
-    ],
-    home_page: 'home-page',
-    site_url: 'https://fake.test',
+    homePage: 'home-page',
+    publicUrl: 'https://fake.test',
     // Specify the theme to use
-    theme_dir: 'test/site/theme',
-    public_dir: 'test/site/theme/public',
-    use_delete_key: true,
-    delete_key: 'test-key',
+    themePath: 'test/site/theme',
+    publicPath: 'test/site/public',
+    useDeleteKey: true,
+    deleteKey: 'test-key',
     plugins: [],
     middleware: [
       ['disable', 'x-powered-by'],