+++ test/tag.test.js
@@ -1,10 +1,10 @@
-const test = require('ava');
-const sinon = require('sinon');
-const request = require('supertest');
+import test from 'ava';
+import sinon from 'sinon';
+import request from 'supertest';
 
-const { UttoriWiki } = require('../src');
+import { UttoriWiki } from '../src/index.js';
 
-const { config, serverSetup, seed } = require('./_helpers/server');
+import { config, serverSetup, seed } from './_helpers/server.js';
 
 test('tagsIndex(request, response, next): renders that tag index page', async (t) => {
   t.plan(3);
@@ -16,7 +16,7 @@ test('tagsIndex(request, response, next): renders that tag index page', async (t
   t.is(response.status, 200);
   t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
   const title = response.text.match(/<title>(.*?)<\/title>/i);
-  t.is(title[1], 'Tags | Wiki');
+  t.is(title[1], 'Tags');
 });
 
 test('tagsIndex(request, response, next): can have middleware set and used', async (t) => {
@@ -64,7 +64,7 @@ test('tag(request, response, next): renders that tag page for a given tag', asyn
   t.is(response.status, 200);
   t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
   const title = response.text.match(/<title>(.*?)<\/title>/i);
-  t.is(title[1], 'Cool | Wiki');
+  t.is(title[1], 'Cool');
 });
 
 test('tag(request, response, next): falls through to 404 when tag is missing', async (t) => {
@@ -77,7 +77,7 @@ test('tag(request, response, next): falls through to 404 when tag is missing', a
   t.is(response.status, 404);
   t.is(response.text.slice(0, 15), '<!DOCTYPE html>');
   const title = response.text.match(/<title>(.*?)<\/title>/i);
-  t.is(title[1], '404 Not Found | Wiki');
+  t.is(title[1], '404 Not Found');
 });
 
 test('tag(request, response, next): can have middleware set and used', async (t) => {